finishing email html

+++ src/controllers/orderController.js
@@ -1,6 +1,8 @@
 import Order from "../models/orderModel.js"
 import Product from "../models/productModel.js"
 import Size from "../models/SizeModel.js"
+import { sendingOrderBYMail } from "./mailsController.js";
+
 
 //function to update product counter
 const updateProductQuantities = async (products, increment) => {
@@ -18,8 +20,10 @@ const updateSizeStock = async (products, increment) => {
     }));
 };
 
+
 export const addOrder = async (req, res) => {
-    const { shippingAddress, status, city, country, totalAmount, deliveryDate, products, userId } = req.body;
+    const { shippingAddress, status, city, country, totalAmount, deliveryDate, products } = req.body;
+    const {userId,email}=req.user;
     let deliveryFee, isFreeDelivery;
 
     if (totalAmount >= 50) {
@@ -30,7 +34,7 @@ export const addOrder = async (req, res) => {
     }
 
     // Check for required fields
-    if (!shippingAddress || !status || !totalAmount || !city || !country || !userId || !products || products.length === 0) {
+    if (!shippingAddress || !totalAmount || !city || !country || !products || products.length === 0) {
         return res.status(400).json({ message: "Missing required field" });
     }
 
@@ -51,7 +55,7 @@ export const addOrder = async (req, res) => {
         // Update product counter and sizes
         await updateProductQuantities(products, 1);
         await updateSizeStock(products, -1);
-
+        sendingOrderBYMail(email,newOrder._id)
         return res.status(201).json({ message: 'Order created successfully', data: newOrder });
     } catch (error) {
         console.log(error);
@@ -75,7 +79,7 @@ export const updateOrder = async (req, res) => {
         const oldOrder = await Order.findById(id);
 
         if (oldOrder.status === "cancelled" || oldOrder.status === "completed") {
-            return res.json({ message: `You can't change the status, order already ${oldOrder.status}`});
+            return res.json({ message: `You can't change the status, order already ${oldOrder.status}` });
         }
 
         if (status === "cancelled") {
@@ -86,7 +90,7 @@ export const updateOrder = async (req, res) => {
 
         if (oldOrder.status === "on-way" || oldOrder.status === "processing") {
             const order = await Order.findByIdAndUpdate(id, { status, products });
-            return res.status(200).json({ message: 'Order updated successfully'});
+            return res.status(200).json({ message: 'Order updated successfully' });
         }
     } catch (error) {
         res.status(500).json({ message: "Error updating order" });
@@ -107,11 +111,17 @@ export const getOneOrderById = async (req, res) => {
 export const deleteOrder = async (req, res) => {
     const { id } = req.params;
     const { products } = req.body;
-
     try {
+        const order = await Order.findById(id)
+        if (!order) {
+            return res.status(400).json({ message: 'There is no order with this id' })
+        }
+
         // Update product counter and stock for the order being deleted
-        await updateProductQuantities(products, -1);
-        await updateSizeStock(products, 1);
+        if (order.status == 'on-way' || order.status == 'processing') {
+            await updateProductQuantities(products, -1);
+            await updateSizeStock(products, 1);
+        }
 
         // Delete the order
         await Order.findByIdAndDelete(id);
